\name{BRCA12}
\docType{data}
\alias{BRCA12}

\title{ BRCA tumour dataset }
\description{
 This data set gives the gene expression values of 30 breast cancer patients.
 Short-term primary fibroblast cultures were established from skin biopsies 
 from 10 BRCA1 and 10 BRCA2 mutation carriers and 10 controls.
}
\usage{BRCA12}
\format{
  \code{ExpressionSet} object containing 8080 genes x 30 pts. 
  Case and controls are specified in \code{Target.class} of \code{phenoData}.
}
\source{The Institute of Cancer Research, Sutton, Surrey, UK}
\references{
 Zsofia Kote-Jarai et al.  \emph{Accurate Prediction of BRCA1 and BRCA2 Heterozygous Genotype Using Expression Profiling after Induced DNA Damage}, Clin Cancer Res 2006;12(13)
}
\keyword{datasets}
