\name{covParams}
\alias{covParams}

\title{ VBMP covariance functions parameters }
\description{
 Returns the value of the covariance functions parameters (theta).
}
\usage{
covParams(obj)
}

\arguments{
  \item{obj}{ an object inheriting from class \code{VBMP.obj}, usually 
    the result of a call to \code{vbmp} }
}
\seealso{ See Also as \code{\link{vbmp}}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }