\name{lowerBound}
\alias{lowerBound}

\title{ VBMP Lower bound estimate }
\description{
 Retruns the lower bound estimates for the VBMP fitted model.
}
\usage{
lowerBound(obj)
}

\arguments{
  \item{obj}{ an object inheriting from class \code{VBMP.obj}, usually 
    the result of a call to \code{vbmp} }
}
\seealso{ See Also as \code{\link{vbmp}}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
