\name{predClass}
\alias{predClass}

\title{ VBMP Predicted class values }
\description{
 Predicted class targets of test dataset.
}
\usage{
predClass(obj)
}

\arguments{
  \item{obj}{ an object inheriting from class \code{VBMP.obj}, usually 
    the result of a call to \code{vbmp} }
}
\seealso{ See Also as \code{\link{vbmp}}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
