\name{predError}
\alias{predError}

\title{ Out-of-Sample VBMP Prediction error }
\description{
 Out-of-Sample Percent Prediction error estimate (0-1 error loss).
}
\usage{
predError(obj)
}

\arguments{
  \item{obj}{ an object inheriting from class \code{VBMP.obj}, usually 
    the result of a call to \code{vbmp} }
}
\seealso{ See Also as \code{\link{vbmp}}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }

